+++ src/entidades/viagem.cpp
@@ -2,7 +2,7 @@
 
 Viagem::~Viagem()
 {
-    for (int i = 0; i < this->destinos_ptr.size(); i++)
+    for (int i = 0; i < static_cast<int>(this->destinos_ptr.size()); i++)
     {
         delete this->destinos_ptr[i];
     }
@@ -11,7 +11,7 @@ Viagem::~Viagem()
 
 int Viagem::get_destino_index(Codigo destino_id)
 {
-    for (int i = 0; i < this->destinos_ptr.size(); i++)
+    for (int i = 0; i < static_cast<int>(this->destinos_ptr.size()); i++)
     {
         if ((this->destinos_ptr[i])->get_codigo() == destino_id)
         {
